restructure project with test directory
