+++ Plot/Program.cs
@@ -1,0 +1,23 @@
+var builder = WebApplication.CreateBuilder(args);
+
+builder.WebHost.UseUrls("http://0.0.0.0:8085");
+
+builder.Services.AddControllers();
+builder.Services.AddOpenApi();
+builder.Services.AddEndpointsApiExplorer();
+
+var app = builder.Build();
+
+if (app.Environment.IsDevelopment())
+{
+    app.MapOpenApi();
+}
+else
+{
+    app.UseHttpsRedirection();
+}
+
+app.MapControllers();
+app.Run();
+
+public partial class Program {};
