--- PlotService.Tests/PlotService_ExampleTest.cs
@@ -1,29 +1,0 @@
-using Xunit;
-using Prime.Services;
-
-namespace Prime.UnitTests.Services
-{
-    public class PrimeService_IsPrimeShould
-    {
-        [Fact]
-        public void IsPrime_InputIs1_ReturnFalse()
-        {
-            var primeService = new PrimeService();
-            bool result = primeService.IsPrime(1);
-
-            Assert.False(result, "1 should not be prime");
-        }
-    }
-
-    public class PrimeService_IsPrimeShouldInverted
-    {
-        [Fact]
-        public void IsPrime_InputIs1_ReturnFalse()
-        {
-            var primeService = new PrimeService();
-            bool result = primeService.IsPrime(1);
-
-            Assert.True(!result, "1 should not be prime");
-        }
-    }
-}
