--- PlotService/PrimeService.cs
@@ -1,16 +1,0 @@
-using System;
-
-namespace Prime.Services
-{
-    public class PrimeService
-    {
-public bool IsPrime(int candidate)
-{
-    if (candidate == 1)
-    {
-        return false;
-    }
-    throw new NotImplementedException("Not fully implemented.");
-}
-    }
-}
