--- PlotService/README.md
@@ -1,1 +1,0 @@
-# PLOT-BE
